refactor: over-engineer solution :)

+++ lib/util.h
@@ -1,3 +1,5 @@
+#include "stdio.h"
+
 /**
  * Wrapper around the normal recv() function.
  * This makes sure that the *buf contains all the bytes sent over by the client.
@@ -5,10 +7,10 @@
  * @param fd the file descriptor of the socket
  * @param buf the buffer to put received data into
  * @param size the maximum allocated size of the buffer
- * @return 0 for success
+ * @return total amount of bytes read
  * @return -1 for failure to read bytes
  */
-int recv_all(int fd, char *buf, long size);
+ssize_t recv_all(int fd, char *buf, long size);
 
 /**
  * Wrapper around the normal send() function.
